LayoutTests/http/tests/security/mixedContent/insecure-script-redirects-to-basic-auth-secure-script-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script-redirects-to-basic-auth-secure-script.html requested insecure content from http://127.0.0.1:8080/resources/redirect.py?url=https://localhost:8443/security/mixedContent/resources/subresource/protected-script.py. This content was blocked and must be served over HTTPS.

This test opens a new window to a secure page that loads an insecure script that redirects to a secure script guarded by basic authentication. The secure script should be blocked because it requires credentials and was loaded via an insecure redirect.

PASS did not load script.
